requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.164
botocore==1.12.164
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
idna==2.8
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
pyaml==19.4.1
Pygments==2.4.0
PyMySQL==0.9.3
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.0
requests==2.22.0
s3fs==0.2.1
s3transfer==0.2.1
scikit-learn==0.21.2
scikit-surprise==1.0.6
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.4
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wincertstore==0.2
